feat(cli): finalize codegen + put metagen and codegen under gen subcommand
